разделение большой картинки на модули - картинка и комментарии, исправление ошибок
